docs(viewer): display multiple viewer in interaction doc
